Move `resolve_dependencies` into resolve.rs
